pyproject.toml
[tool.poetry]
name = "homework8"
version = "1.0.0"
description = ""
authors = ["Riabets Gleb <work.ga@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
pydantic = {extras = ["email"], version = "^1.9.0"}
fastapi = "^0.75.0"
uvicorn = "^0.17.6"
redis = "^4.2.2"
charset-normalizer = "^2.0.12"
types-redis = "^4.1.22"
types-requests = "^2.27.20"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.36"}
asyncpg = "^0.25.0"
aiosqlite = "^0.17.0"
fastapi-jwt-auth = "^0.5.0"
passlib = "^1.7.4"
asyncclick = "^8.0.3"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.18.3"
pytest-deadfixtures = "^2.2.1"
pytest-cov = "^3.0.0"
pytest-mock = "^3.7.0"
autoflake = "^1.4"
isort = "^5.10.0"
black = "^22.1"
unify = "^0.5"
flake8 = "^4.0"
pylint = "^2.12.1"
mypy = "^0.940"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
